+++ tests/__init__.py
@@ -32,22 +32,54 @@ class TestNk3Updates(unittest.TestCase):
             _Migration.get(Variant.NRF52, Version(1, 0, 0), Version(1, 1, 0)),
             frozenset(),
         )
+        self.assertEqual(
+            _Migration.get(None, Version(1, 8, 2), Version(1, 9, 0)),
+            frozenset(),
+        )
 
-    def test_update_path_match(self) -> None:
+    def test_update_path_nrf(self) -> None:
         from nitrokey.nk3.updates import _Migration
         from nitrokey.trussed import Variant, Version
 
-        nrf_migration = frozenset([_Migration.NRF_IFS_MIGRATION])
+        migrations = frozenset([_Migration.NRF_IFS_MIGRATION])
 
         self.assertEqual(
             _Migration.get(Variant.NRF52, Version(1, 2, 2), Version(1, 3, 0)),
-            nrf_migration,
+            migrations,
         )
         self.assertEqual(
             _Migration.get(Variant.NRF52, Version(1, 0, 0), Version(1, 3, 0)),
-            nrf_migration,
+            migrations,
         )
         self.assertEqual(
             _Migration.get(Variant.NRF52, None, Version(1, 3, 0)),
-            nrf_migration,
+            migrations,
+        )
+
+    def test_update_path_ifs_v2(self) -> None:
+        from nitrokey.nk3.updates import _Migration
+        from nitrokey.trussed import Variant, Version
+
+        migrations = frozenset([_Migration.IFS_MIGRATION_V2])
+
+        self.assertEqual(
+            _Migration.get(Variant.NRF52, Version(1, 5, 0), Version(1, 8, 2)),
+            migrations,
+        )
+        self.assertEqual(
+            _Migration.get(None, Version(1, 8, 0), Version(1, 8, 2)),
+            migrations,
+        )
+
+    def test_update_path_multi(self) -> None:
+        from nitrokey.nk3.updates import _Migration
+        from nitrokey.trussed import Variant, Version
+
+        migrations = frozenset(
+            [_Migration.NRF_IFS_MIGRATION, _Migration.IFS_MIGRATION_V2]
+        )
+
+        self.assertEqual(
+            _Migration.get(Variant.NRF52, Version(1, 2, 2), Version(1, 8, 2)),
+            migrations,
         )